TTT: types, templates, and tenacity

+++ Framework/include/BJet.h
@@ -22,8 +22,8 @@ private:
     void bjet(NTupleReader& tr)
     {
         const auto& Jets = tr.getVec<utility::LorentzVector>("Jets"+myVarSuffix_);
-        const auto& Jets_bJetTagDeepCSVtotb = tr.getVec<double>("Jets"+myVarSuffix_+"_bJetTagDeepCSVtotb");
-        const auto& Jets_bDiscriminatorCSV = tr.getVec<double>("Jets"+myVarSuffix_+"_bDiscriminatorCSV"); //
+        const auto& Jets_bJetTagDeepCSVtotb = tr.getVec<float>("Jets"+myVarSuffix_+"_bJetTagDeepCSVtotb");
+        const auto& Jets_bDiscriminatorCSV = tr.getVec<float>("Jets"+myVarSuffix_+"_bDiscriminatorCSV"); //
         const auto& etaCut = tr.getVar<double>("etaCut");
         const auto& JetsID = tr.getVec<bool>("Jets"+myVarSuffix_+"_ID");
         const auto& GoodJets = tr.getVec<bool>("GoodJets"+myVarSuffix_);
